refactor: add interceptor to the rest service
