+++ src/app/app.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import {HttpClientModule, HttpHandler} from '@angular/common/http';
+import {HTTP_INTERCEPTORS, HttpClientModule, HttpHandler} from '@angular/common/http';
 import { ReactiveFormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import {RestService} from "./services/rest/rest.service";
@@ -28,6 +28,10 @@ import {RouteService} from "./services/route/route.service";
     RestService,
     AuthService,
     RouteService,
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: RestService,
+      multi: true },
     {
       provide: RestService,
       useFactory: restServiceFactory,
